+++ plugins/postcss-progressive-custom-properties/generate/matchers.mjs
@@ -1,4 +1,5 @@
 import { promises as fsp } from 'fs';
+import { colorMixMatchers } from './color-mix.mjs';
 import { colorMatchers, hslMatchers, hwbMatchers, labMatchers, lchMatchers, oklabMatchers, oklchMatchers, rgbMatchers } from './color.mjs';
 import { icUnitMatchers } from './font-size.mjs';
 
@@ -15,6 +16,9 @@ fsp.writeFile(
 			...oklabMatchers,
 			...oklchMatchers,
 			...rgbMatchers,
+
+			// color mix:
+			...colorMixMatchers,
 
 			// font-size:
 			...icUnitMatchers,